sched : refactor scheduler feature
Integration of distributed scheduler files in the guest.
If you are developing a scheduler, at that time,
you can create file.(scheduler, policy...)
