--- hypervisor/sched_policy.h
@@ -1,8 +1,0 @@
-#ifndef __SCHED_POLICY_H__
-#define __SCHED_POLICY_H__
-
-#include <hvmm_types.h>
-
-vmid_t sched_policy_determ_next(void);
-
-#endif
