+++ platform-device/cortex_a15x2_arndale/Makefile
@@ -32,11 +32,10 @@ OBJS 		= boot.o	\
 	$(HYPERVISOR_SOURCE_DIR)/gic/gic.o				\
 	$(HYPERVISOR_SOURCE_DIR)/interrupt.o			\
 	$(HYPERVISOR_SOURCE_DIR)/smp.o					\
-	$(HYPERVISOR_SOURCE_DIR)/scheduler.o			\
 	$(HYPERVISOR_SOURCE_DIR)/generic_timer/generic_timer.o	\
 	$(HYPERVISOR_SOURCE_DIR)/gic/vgic.o				\
 	$(HYPERVISOR_SOURCE_DIR)/trap.o					\
-	$(HYPERVISOR_SOURCE_DIR)/sched_policy.o			\
+	$(HYPERVISOR_SOURCE_DIR)/guest.o				\
 	$(HYPERVISOR_SOURCE_DIR)/trap_dabort.o			\
 	$(HYPERVISOR_SOURCE_DIR)/vdev/vdev.o			\
 	$(HYPERVISOR_SOURCE_DIR)/vdev/vdev_sample.o		\
